move Pipeline to root path
